feat: add Prettier and ESLint configuration files, update component structure, and refactor code formatting
