+++ dashboard/.eslintrc.json
@@ -1,0 +1,14 @@
+{
+    "extends": [
+        "eslint:recommended",
+        "next/core-web-vitals",
+        "prettier",
+        "plugin:cypress/recommended"
+    ],
+    "plugins": ["unused-imports"],
+    "rules": {
+        "unused-imports/no-unused-imports": "error",
+        "react-hooks/rules-of-hooks": "error",
+        "react-hooks/exhaustive-deps": "off"
+    }
+}
