+++ dashboard/.prettierrc
@@ -1,0 +1,7 @@
+{
+    "printWidth": 80,
+    "trailingComma": "none",
+    "tabWidth": 4,
+    "semi": true,
+    "singleQuote": false
+}
