+++ dashboard/next.config.ts
@@ -1,14 +1,14 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "**", // Allows all hosts
-      },
-    ],
-  },
+    images: {
+        remotePatterns: [
+            {
+                protocol: "https",
+                hostname: "**" // Allows all hosts
+            }
+        ]
+    }
 };
 
 export default nextConfig;
