+++ dashboard/src/app/order/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function page() {
-  return <div>This is order page</div>;
+    return <div>This is order page</div>;
 }
