+++ dashboard/src/app/page.tsx
@@ -1,7 +1,7 @@
 export default function Home() {
-  return (
-    <div>
-      <p>This is Homepage</p>
-    </div>
-  );
+    return (
+        <div>
+            <p>This is Homepage</p>
+        </div>
+    );
 }
