+++ dashboard/src/app/user/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function page() {
-  return <div>This is user page</div>;
+    return <div>This is user page</div>;
 }
